migrate /@/providers/stampPicker to pinia
