+++ src/components/Main/MainView/MessageElement/MessageStampList.vue
@@ -46,7 +46,7 @@ import { StampId, UserId } from '/@/types/entity-ids'
 import StampDetailElement from './StampDetailElement.vue'
 import AIcon from '/@/components/UI/AIcon.vue'
 import apis from '/@/lib/apis'
-import { useStampPickerInvoker } from '/@/providers/stampPicker'
+import { useStampPickerInvoker } from '/@/store/ui/stampPicker'
 import { useToastStore } from '/@/store/ui/toast'
 import { useMeStore } from '/@/store/domain/me'
 import { useStampsStore } from '/@/store/entities/stamps'
