+++ src/components/Main/MainView/MessageElement/MessageTools.vue
@@ -85,7 +85,7 @@ import { defineComponent, computed, PropType, ref } from 'vue'
 import AIcon from '/@/components/UI/AIcon.vue'
 import AStamp from '/@/components/UI/AStamp.vue'
 import { StampId, MessageId } from '/@/types/entity-ids'
-import { useStampPickerInvoker } from '/@/providers/stampPicker'
+import { useStampPickerInvoker } from '/@/store/ui/stampPicker'
 import { useResponsiveStore } from '/@/store/ui/responsive'
 import apis from '/@/lib/apis'
 import { useToastStore } from '/@/store/ui/toast'
